% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-state.R
\name{allow_solution_error}
\alias{allow_solution_error}
\title{Set a flag allowing student code to produce an error.}
\usage{
allow_solution_error()
}
\description{
When your exercise is supposed to produce an error, include this function in 
your pre-exercise code to set a flag that allows the code to return the state
without error-ing out. **Disclaimer**: this is just an attempt to emulate
what would happen if you did the same thing in a DataCamp exercise, but we
have no idea if it that how DataCamp actually accomplishes this.
}
