% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-rmd.R
\name{check_equal.MarkdownChunkOptionState}
\alias{check_equal.MarkdownChunkOptionState}
\title{Check equality of markdown code chunk option}
\usage{
\method{check_equal}{MarkdownChunkOptionState}(state,
  incorrect_msg = NULL, append = FALSE, ...)
}
\arguments{
\item{state}{the state to start from. Should be a state produced by \code{\link{check_option}}.}

\item{incorrect_msg}{If specified, this overrides the automatically generated message in case the options don't match between student and solution.}

\item{append}{Whether or not to append the feedback to feedback built in previous states.}

\item{...}{S3 stuff}
}
\description{
Check equality of markdown code chunk option
}
