% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-error.R
\name{check_error}
\alias{check_error}
\alias{check_error.default}
\title{Check whether the student's submission threw an error.}
\usage{
check_error(state, ...)

\method{check_error}{default}(state, incorrect_msg = NULL, ...)
}
\arguments{
\item{state}{State to start from (for \code{check_error})}

\item{...}{S3 stuff}

\item{incorrect_msg}{additional message that is appended to the automatically
generated feedback message.}
}
\description{
With information gathered from the R Backend, detect whether the student's
submission generated an error. If this is the case a feedback message will be
generated that contains the R error in a formatted style. You can use
\code{incorrect_msg} to add a custom message to this error.
}
\examples{
\dontrun{
# Example student code: x <- 4 + "a"

# SCT
ex() \%>\% check_error()
}

}
