% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-file.R
\name{check_file.FileSysState}
\alias{check_file.FileSysState}
\title{Try to get code from files submitted by student.}
\usage{
\method{check_file}{FileSysState}(state, name, missing_msg = NULL)
}
\arguments{
\item{state}{a}

\item{name}{full file name (e.g. script.R)}

\item{missing_msg}{feedback to give if file is missing}
}
\description{
Try to get code from files submitted by student.
}
