% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-function-result.R
\name{check_function_result}
\alias{check_function_result}
\alias{check_result.OperationState}
\alias{check_result.FunctionState}
\alias{check_equal.FunctionResultState}
\alias{check_equal.OperationResultState}
\title{Check the result of a function call/operation}
\usage{
\method{check_result}{OperationState}(state, error_msg = NULL,
  append = TRUE, ...)

\method{check_result}{FunctionState}(state, error_msg = NULL,
  append = TRUE, ...)

\method{check_equal}{FunctionResultState}(state,
  eq_condition = "equivalent", eq_fun = NULL, incorrect_msg = NULL,
  append = TRUE, ...)

\method{check_equal}{OperationResultState}(state,
  eq_condition = "equivalent", eq_fun = NULL, incorrect_msg = NULL,
  append = TRUE, ...)
}
\arguments{
\item{state}{the state to start from (for \code{check_} functions)}

\item{error_msg}{feedback message in case the student function call at the
mentioned index generated an error.}

\item{append}{Whether or not to append the feedback to feedback built in
previous states}

\item{...}{S3 stuff}

\item{eq_condition}{character string indicating how to compare. See
\code{\link{is_equal}}.}

\item{eq_fun}{optional argument to specify a custom equality function. The
function should take two arguments and always return a single boolean
value: \code{TRUE} or \code{FALSE}.}

\item{incorrect_msg}{feedback message in case the evaluation was not the
same as in the solution.}
}
\description{
Check the result of a function call/operation
}
\examples{
\dontrun{
# Example 1
mean(1:3)

# SCT
ex() \%>\% check_function("mean") \%>\% check_result() \%>\% check_equal()
}
}
