% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-ggplot.R
\name{check_ggplot}
\alias{check_ggplot}
\title{Test ggplot call}
\usage{
check_ggplot(state, index = 1, all_fail_msg = NULL, check_data = TRUE,
  data_fail_msg = NULL, check_aes = TRUE, aes_fail_msg = NULL,
  exact_aes = FALSE, check_geom = TRUE, geom_fail_msg = NULL,
  exact_geom = FALSE, check_geom_params = NULL, check_facet = TRUE,
  facet_fail_msg = NULL, check_scale = TRUE, scale_fail_msg = NULL,
  exact_scale = FALSE, check_coord = TRUE, coord_fail_msg = NULL,
  exact_coord = FALSE, check_stat = TRUE, stat_fail_msg = NULL,
  exact_stat = FALSE, check_extra = NULL, extra_fail_msg = NULL,
  exact_extra = NULL, check = NULL)
}
\arguments{
\item{state}{the state to start from}

\item{index}{which call to check}

\item{all_fail_msg}{Message if all fails}

\item{check_data}{Whether or not to check data layer}

\item{data_fail_msg}{Message in case data layer fails}

\item{check_aes}{Whether or not to check aes layer}

\item{aes_fail_msg}{Message in case aes layer fails}

\item{exact_aes}{Should the aesthetics be exact?}

\item{check_geom}{Whether or not to check geom layer}

\item{geom_fail_msg}{Message in case geom layer fails}

\item{exact_geom}{Should the geoms be exact?}

\item{check_geom_params}{Should the geom parameters be checked?}

\item{check_facet}{Whether or not to check facet layer}

\item{facet_fail_msg}{Message in case facet layer fails}

\item{check_scale}{Whether or not to check scale layer}

\item{scale_fail_msg}{Message in case scale layer fails}

\item{exact_scale}{Whether or not scales should be defined exactly}

\item{check_coord}{Whether or not to check coord layer}

\item{coord_fail_msg}{Message in case coord layer fails}

\item{exact_coord}{Whether or not coords should be defined exactly}

\item{check_stat}{Whether or not to check stat layer}

\item{stat_fail_msg}{Message in case stat layer fails}

\item{exact_stat}{Whether or not stats should be defined exactly}

\item{check_extra}{Whether to check extra stuff}

\item{extra_fail_msg}{Message in case extra stuff fails}

\item{exact_extra}{Whether or not extra info should be exactly specified.}

\item{check}{Which layers to check}
}
\description{
Test ggplot call
}
