% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-description.R
\name{check_has_desc_element}
\alias{check_has_desc_element}
\title{Check the DESCRIPTION fields}
\usage{
check_has_desc_element(state, element, missing_msg = NULL, append = TRUE)
}
\arguments{
\item{state}{The state of the exercise, as returned from \code{\link{parse_desc}}.}

\item{element}{String naming the element of the roxygen block to check.}

\item{missing_msg}{Optional string. Used to override the feedback message
in the event of failure.}

\item{append}{For compatibility; currently unused.}

\item{regex}{String providing a regular expression for the solution code to 
match. See \code{\link{check_code}}.}

\item{fixed}{Logical. If \code{TRUE}, regex is treated as a fixed string, not
a regular expression. See \code{\link{check_code}}.}

\item{times}{Positive integer. Denotes the number of times the string in 
\code{regex} should be matched.}

\item{incorrect_msg}{Optional string. Used to override the feedback message
in the event of failure.}

\item{not_typed_msg}{Optional string. Used to override the feedback message
in the event of failure.
incorrect_msg}
}
\value{
This function is invoked for the side effect of registering feedback 
in the event of a failed test. See \code{\link{check_that}} for details of
the return value and feedback mechanism.
\code{check_has_desc_element} checks that the \code{element} element of the 
DESCRIPTION is present.
\code{check_desc_element_equals} checks that the \code{element} element of 
the DESCRIPTION is equal to the value in the solution code.
\code{check_desc_element_matches} checks that the \code{element} element of 
the DESCRIPTION matches a regular expression or string.
}
\description{
Check that the DESCRIPTION fields provided by the student are correct.
}
\examples{
\dontrun{
  # Always begin by calling parse_desc() on the exercise state.
  ex() \%>\% parse_roxy() \%>\% {
    check_has_desc_element(., 'title')
    check_roxy_element_equals(., 'version', as.numeric_version('1.0.0'))
    check_roxy_element_matches(., 'description', 'data.+manipulation')
  }
}
}
