% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-roxygen.R
\name{check_has_roxy}
\alias{check_has_roxy}
\alias{check_has_roxy_element}
\title{Check the roxygen comments}
\usage{
check_has_roxy(state, index = 1L, missing_msg = NULL, append = TRUE)

check_has_roxy_element(state, element, index = 1L, missing_msg = NULL,
  append = TRUE)
}
\arguments{
\item{state}{The state of the exercise, as returned from \code{\link{parse_roxy}}.}

\item{index}{A positive integer or a string naming a function. This describes 
which roxygen element in the code to check.}

\item{missing_msg}{Optional string. Used to override the feedback message
in the event of failure.}

\item{append}{For compatibility; currently unused.}

\item{element}{String naming the element of the roxygen block to check.}

\item{incorrect_msg}{Optional string. Used to override the feedback message
in the event of failure.}

\item{not_typed_msg}{Optional string. Used to override the feedback message
in the event of failure.
incorrect_msg}
}
\value{
This function is invoked for the side effect of registering feedback 
in the event of a failed test. See \code{\link{check_that}} for details of
the return value and feedback mechanism.
}
\description{
Check that the roxygen comments provided by the student are correct.
}
\details{
\code{check_has_roxy} checks that the \code{index} block of roxygen
is present.
\code{check_has_roxy_element} checks that the \code{element} element of the 
\code{index} block of roxygen is present.
\code{check_roxy_element_equals} checks that the \code{element} element of 
the \code{index} block of roxygen is equal to the value in the solution code.
\code{check_roxy_element_matches} checks that the \code{element} element of 
the \code{index} block of roxygen matches a regular expression or string.
}
\examples{
\dontrun{
  # Always begin by calling parse_roxy() on the exercise state.
  ex() \%>\% parse_roxy() \%>\% {
    check_has_roxy(., index = 2)
    check_has_roxy_element(., 'title', index = 2)
    check_roxy_element_equals(., 'description', 'This is a mean function')
    check_roxy_element_matches(., 'return', 'integer +vector')
  }
}
}
