% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-that.R
\name{check_that}
\alias{check_that}
\title{Expectation wrapper}
\usage{
check_that(code, feedback)
}
\arguments{
\item{code}{The expectation that should be wrapped}

\item{feedback}{A character string with feedback when the expection is not
met OR a list object, containing multiple pieces of information. This list
should at least contain an element named \code{message}}
}
\description{
This function wraps around an is_... function. When the expectation fails to
be met, the feedback message is sent to the reporter. You can use 
\code{\link{is_true}}, \code{\link{is_false}}, \code{\link{is_gte}} or
\code{\link{is_equal}}
}
\examples{
\dontrun{
check_that(is_true(3 == 3))
check_that(is_false(3 == 4))
check_that(is_gte(4, 3))
check_that(is_equal(4, 4))
}

}

