% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{compare}
\alias{compare}
\alias{is_true}
\alias{is_false}
\alias{is_gte}
\alias{is_lt}
\title{Check if object is true, false or >=}
\usage{
is_true(x)

is_false(x)

is_gte(x, y)

is_lt(x, y)
}
\arguments{
\item{x}{object to test}

\item{y}{single numeric value to compare}
}
\description{
Utility functions to use inside \code{\link{check_that}}.
}
