% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reporter-DataCamp.R
\docType{class}
\name{DataCampReporter-class}
\alias{DataCampReporter}
\alias{DataCampReporter-class}
\title{DataCamp reporter: gather test results along with elapsed time and
feedback messages.}
\description{
This reporter gathers all results, adding additional information such as
test elapsed time and feedback messages.
}
\keyword{debugging}

