% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-DataCamp.R
\docType{data}
\name{DataCampReporter}
\alias{DataCampReporter}
\title{DataCamp reporter: gather test results along with elapsed time and
feedback messages.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataCampReporter
}
\description{
This reporter gathers all results, adding additional information such as
test elapsed time and feedback messages.
}
\keyword{debugging}

