% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-DataCamp.R
\docType{data}
\name{DC_reporter}
\alias{DC_reporter}
\title{DataCamp reporter: 'log' test results}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DC_reporter
}
\description{
DataCamp reporter: 'log' test results
}
\keyword{datasets}

