% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{disable_highlighting}
\alias{disable_highlighting}
\title{Disable highlighting for any future checks in the chain}
\usage{
disable_highlighting(state)
}
\arguments{
\item{state}{the state to create a substate from}
}
\description{
If the function is used right after \code{ex()}, highlighting is disabled for
the rest of the chain
}
\details{
If the function used 'deep in a chain', the system will fall back on
highlights that were collected earlier in the state.
}
