% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_solution}
\alias{execute_solution}
\title{Execute code in the solution environment}
\usage{
execute_solution(code)
}
\arguments{
\item{code}{The code to execute}
}
\description{
Execute code in the solution environment
}
