% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{execute_student}
\alias{execute_student}
\title{Execute code in the student environment}
\usage{
execute_student(code)
}
\arguments{
\item{code}{The code to execute}
}
\description{
Execute code in the student environment
}
