% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{extract_roxygen_from_code}
\alias{extract_roxygen_from_code}
\title{Extract roxygen details from a file}
\usage{
extract_roxygen_from_code(lines)
}
\arguments{
\item{file}{A string denoting a path to an R file.}
}
\value{
A list of lists. Each top level element corresponds to a roxygen 
block. Each second level element corresponds to a roxygen tag within that 
block.
}
\description{
Parses an R file and extracts the roxygen tags. Mostly just a wrapper
around \code{roxygen2:::parse_blocks}.
}
