% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-old.R
\name{get_clean_lines}
\alias{get_clean_lines}
\title{convert student/solution code to vector of clean strings with the pipe operator removed.}
\usage{
get_clean_lines(code)
}
\arguments{
\item{code}{the code to convert to a vector of unpiped clean strings}
}
\description{
convert student/solution code to vector of clean strings with the pipe operator removed.
}

