% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_hits}
\alias{get_num_hits}
\title{Get the number of hits for a series of regexes}
\usage{
get_num_hits(regex, x, fixed)
}
\arguments{
\item{regex}{vector of regular expressions against which to match}

\item{x}{character vector where matches are sought}

\item{fixed}{logical. If \code{TRUE}, \code{regex} are strings to be matched as is.}
}
\description{
Get the number of hits for a series of regexes
}
