% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_tags}
\alias{init_tags}
\title{(re)initialize the tags for logging}
\usage{
init_tags(...)
}
\arguments{
\item{...}{The tags you want to set, e.g. \code{fun = "test_object"}}
}
\description{
All previous tags that have been set previously will be cleared.
}

