% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-that.R
\name{is_equal}
\alias{is_equal}
\title{Check equality of two objects}
\usage{
is_equal(x, y, eq_condition = "equivalent")
}
\arguments{
\item{x}{object to test}

\item{y}{object to compare}

\item{eq_condition}{how to compare the objects: \code{"equivalent"} (the default,
does not check attributes), \code{"equal"} (checks attributes, but allows for
errors in machine precision), or \code{"identical"} (exactly identical).}
}
\description{
Utility function to use inside \code{\link{check_that}}.
}

