% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-that.R
\name{is_false}
\alias{is_false}
\title{Check if object is false	
	
Utility function to use inside \code{\link{check_that}}.}
\usage{
is_false(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Check if object is false	
	
Utility function to use inside \code{\link{check_that}}.
}
\keyword{internal}
