% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-that.R
\name{is_gte}
\alias{is_gte}
\title{Check >= relation	
	
Utility function to use inside \code{\link{check_that}}.}
\usage{
is_gte(x, y)
}
\arguments{
\item{x}{object to test}

\item{y}{single numeric value to compare}
}
\description{
Check >= relation	
	
Utility function to use inside \code{\link{check_that}}.
}
\keyword{internal}
