% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{override}
\alias{override}
\alias{override_solution}
\alias{override_solution_code}
\alias{override_solution_env}
\title{Functions to override solution code (and parse data) and variables in the solution environment.}
\usage{
override_solution(state, code = NULL, ...)

override_solution_code(state, code)

override_solution_env(state, ...)
}
\arguments{
\item{state}{the state to create a substate from}

\item{code}{the solution code to put into the state}

\item{...}{named environment variables to add to or override in the solution environment}
}
\description{
Produces a new state with a custom solution code. Mostly useful inside
\code{test_or}, if you want to test for different cases.
}
