% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{override_solution}
\alias{override_solution}
\title{Override the solution code and pd in a state}
\usage{
override_solution(state, code)
}
\arguments{
\item{state}{the state to create a substate from}

\item{code}{the solution code to put into the state}
}
\description{
Produces a new state with a custom solution code. Mostly useful inside
\code{test_or}, if you want to test for different cases.
}

