% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api-overrides.R
\name{overrides}
\alias{get_sct_result}
\alias{get_solution_code}
\alias{get_solution_env}
\alias{get_student_code}
\alias{get_student_error}
\alias{get_student_output}
\alias{overrides}
\alias{set_sct_result}
\alias{set_solution_code}
\alias{set_student_code}
\alias{set_student_error}
\alias{set_student_output}
\title{DataCamp API overrides}
\usage{
get_solution_env()

get_student_code()

set_student_code(x)

get_solution_code()

set_solution_code(x)

get_student_output()

set_student_output(x)

get_student_error()

set_student_error(x)

get_sct_result()

set_sct_result(x)
}
\description{
Functions that override the API functionality if it is not available.
}
\keyword{internal}

