% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-rcpp.R
\name{parse_rcpp}
\alias{parse_rcpp}
\title{Parse Rcpp code}
\usage{
parse_rcpp(state)
}
\arguments{
\item{state}{An exercise state, as returned by \code{ex()}.}
}
\value{
A child state.
}
\description{
Parses Rcpp code (currently just the R chunks) and updates the state.
}
\details{
The function extracts the R code lines from the state then parses
them. Currently the C++ code lines are not parsed, so they can only be 
checked using \code{check_code()}-based SCTs.
}
\keyword{internal}
