% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-exercise.R
\name{run_until_fail}
\alias{run_until_fail}
\title{Run SCT until it fails}
\usage{
run_until_fail(code, envir = NULL)
}
\arguments{
\item{code}{the SCT script to run as an expression}

\item{envir}{environment in which to execute the SCT script}
}
\description{
Run SCT until it fails
}
