% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-definitions.R
\name{s3definitions}
\alias{s3definitions}
\alias{check_equal}
\alias{check_equal.default}
\alias{check_result}
\alias{check_output}
\alias{check_error}
\alias{check_body}
\alias{check_body.default}
\alias{check_file}
\alias{check_file.default}
\title{S3 definitions}
\usage{
check_equal(state, ...)

\method{check_equal}{default}(state, ...)

check_result(state, ...)

check_output(state, ...)

check_error(state, ...)

check_body(state, ...)

\method{check_body}{default}(state, ...)

check_file(state, ...)

\method{check_file}{default}(state, ...)
}
\arguments{
\item{state}{state to start from}

\item{...}{arguments passed to S3 implementations}
}
\description{
Functions that are extended to fit different cases.
}
