% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages-utils.R
\name{set_language}
\alias{set_language}
\title{Specify the language for the automated feedback}
\usage{
set_language(lang)
}
\arguments{
\item{lang}{shorthand notation for the language.}
}
\description{
Functionality for different languages is no longer supported
}
