% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_tags}
\alias{set_tags}
\title{Set the tags for logging}
\usage{
set_tags(...)
}
\arguments{
\item{...}{The tags you want to set, e.g. \code{fun = "test_object"}}
}
\description{
Set the tags for logging
}

