% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-state.R
\name{setup_state}
\alias{setup_state}
\title{Set up state for local experimentation.}
\usage{
setup_state(sol_code = "", stu_code = "", sol_env = NULL,
  stu_env = NULL, stu_result = NULL, pec = character(),
  ex_type = "NormalExercise")
}
\arguments{
\item{sol_code}{Solution script as a string. If it is not specified, the
student code will be used.}

\item{stu_code}{Student submission as a string. If it is not specified, the
solution code will be used.}

\item{sol_env}{Solution environment. If this is specified, the solution code is not rerun.}

\item{stu_env}{Student environment. If this is specified, the student code is not rerun.}

\item{stu_result}{Result of calling \code{\link{evaluate}} on the student code. If this is
is specified, this overrides the output generated by running \code{stu-code}.}

\item{pec}{Pre-exercise-code as a string}

\item{ex_type}{Type of exercise as a string. Defaults to NormalExercise.}
}
\value{
The exercise state, from which you can start chaining.
}
\description{
It runs both the solution and the student submission, and populates the state
with parse data, output, etc. After running this function, the state is
available thorugh \code{\link{ex}}, from which you can start your SCT chains.
In a way, this function is a very light weight version of DataCamp's R Backend.
}
\note{
This function is only supposed to be used locally when experimenting.
  It should never be used in the eventual SCT script of an exercise.
}
\examples{
\dontrun{
setup_state(
 sol_code = "a <- 1",
 stu_code = "a <- 2"
)

ex() \%>\% check_object('a') \%>\% check_equal()
}

}
