% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{state}
\alias{state}
\alias{State}
\alias{RootState}
\alias{ChildState}
\title{testwhat states.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
State

RootState

ChildState
}
\description{
Root State has no parent state.
ChildState does have state.
Both inherit from the prototypical State class
}
\keyword{datasets}
