% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-arg.R
\name{test_arg}
\alias{test_arg}
\title{Test if student specified argument}
\usage{
test_arg(state, arg, arg_not_specified_msg = NULL)
}
\arguments{
\item{arg}{name of argument to specify}

\item{arg_not_specified_msg}{custom message in case argument was not specified}
}
\description{
Test if student specified argument
}

