% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-test-axis.R
\name{test_axis}
\alias{test_axis}
\title{Test whether the student correctly defined axis properties (ggvis exercises)}
\usage{
test_axis(index = 1, type = NULL, props = NULL, not_called_msg = NULL,
  incorrect_msg = NULL, incorrect_number_of_calls_msg = NULL)
}
\arguments{
\item{index}{exercise to be checked (solution and student code should have same number of calls!)
properties inside the first mentioned function by the teacher.}

\item{type}{which axis to check (x or y). Only one axis is tested at the same time.}

\item{props}{set of axis properties to be checked. If not specified, all properties found in the solution or checked on.}

\item{not_called_msg}{feedback message in case the specified axis type (x or y) was not found}

\item{incorrect_msg}{feedback message in case the axes properties defined by the student did not correspond with the one of the solution.}

\item{incorrect_number_of_calls_msg}{feedback message in case the student did
enter the same amount of commands as the solution did.}
}
\description{
Test whether the student correctly assigned axis properties. The student's and solution code is automatically compared to
one another for specific properties of the axes.
}

