% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-body.R
\name{test_body}
\alias{test_body}
\title{Test if student coded body correctly}
\usage{
test_body(state)
}
\arguments{
\item{state}{state to start from}
}
\description{
body can be of for, while or function definition
}

