% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-call.R
\name{test_call}
\alias{test_call}
\alias{test_fun}
\alias{test_op}
\title{Test if user called function/operator}
\usage{
test_fun(state, name, index = 1, not_called_msg = NULL)

test_op(state, name, index = 1, not_called_msg = NULL)
}
\arguments{
\item{state}{state to start from}

\item{name}{name of function/function as a string, e.g. \code{"mean"} or \code{"+"}}

\item{index}{which call of \code{name} in the solution to check.}

\item{not_called_msg}{custom message in case \code{name} was not called sufficiently}
}
\description{
Test if user called function/operator
}

