% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-call-result.R
\name{test_call_result}
\alias{check_equal.FunctionResultState}
\alias{check_equal.OperationResultState}
\alias{check_result.FunctionState}
\alias{check_result.OperationState}
\alias{test_call_result}
\alias{test_function_result}
\title{Check the result of a function call/operation}
\usage{
test_function_result(name = NULL, index = 1, eq_condition = "equivalent",
  not_called_msg = NULL, error_msg = NULL, incorrect_msg = NULL)

\method{check_result}{OperationState}(state, error_msg = NULL,
  append = TRUE, ...)

\method{check_result}{FunctionState}(state, error_msg = NULL, append = TRUE,
  ...)

\method{check_equal}{FunctionResultState}(state, eq_condition = "equivalent",
  incorrect_msg = NULL, append = TRUE, ...)

\method{check_equal}{OperationResultState}(state, eq_condition = "equivalent",
  incorrect_msg = NULL, append = TRUE, ...)
}
\arguments{
\item{name}{name of the function whose output you would like to check.}

\item{index}{Ordinal number of the call you want to check (both student and solution!).}

\item{eq_condition}{how to compare student and solution.}

\item{not_called_msg}{feedback message in case the function is not retrieved.}

\item{error_msg}{feedback message in case the student function call at the mentioned index generated an error.}

\item{incorrect_msg}{feedback message in case the evaluation was not the same as in the solution.}

\item{state}{the state to start from (for \code{check_} functions)}

\item{append}{Whether or not to append the feedback to feedback built in previous states}

\item{...}{S3 stuff}
}
\description{
Check the result of a function call/operation
}
\examples{
\dontrun{
# Example 1
mean(1:3)

# SCT option 1
test_function_result("mean")

# SCT option 2
ex() \%>\% check_function("mean") \%>\% check_result() \%>\% check_equal()
}
}

