% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-code.R
\name{test_code}
\alias{check_code}
\alias{test_code}
\alias{test_student_typed}
\title{Test the student's code as text}
\usage{
test_student_typed(strings, fixed = TRUE, times = 1, not_typed_msg = NULL)

check_code(state, regex, fixed = FALSE, times = 1, missing_msg = NULL,
  append = TRUE)
}
\arguments{
\item{strings}{A set of strings/regexes that should be in the student code.}

\item{fixed}{if TRUE, strings are treated literally. If FALSE, strings are
treated as regex patterns.}

\item{times}{how often should any of the strings be matched?}

\item{not_typed_msg}{Custom feedback in case the pattern is not contained
often enough in the student's submission.}

\item{state}{the state to start from}

\item{regex}{A set of strings/regexes that should be in the student code.}

\item{missing_msg}{Custom feedback in case the pattern is not contained often
enough in the student's submission.}

\item{append}{Whether or not to append the feedback to feedback built in previous states}
}
\description{
Some rudimentary string cleaning is performed to allow for different ways of
saying the same things (removing spaces, changing single quotes to double
quotes, changing TRUE to T ...).
}
\details{
Using these function should be a last resort, as there are myriad ways of 
solving the same problems in R!

Watch out: in \code{test_student_typed} the default is \code{fixed = TRUE},
in \code{check_code} the default is \code{fixed = FALSE}.
}
\examples{
\dontrun{
# Example 1
TRUE & FALSE

# SCT option 1
test_student_typed(c("TRUE & FALSE", "FALSE & TRUE"))

# SCT option 2
ex() \%>\% check_code(c("TRUE & FALSE", "FALSE & TRUE"), fixed = TRUE)

# Example 2:
"Hello, world!"

# SCT option 1, robust to small typos
test_student_typed("[H|h]ello,*\\\\s*[W|w]orld\\\\!*", fixed = FALSE)

# SCT option 2, robust to small typos
ex() \%>\% check_code("[H|h]ello,*\\\\s*[W|w]orld\\\\!*")
}

}

