% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-control.R
\name{test_control}
\alias{test_control}
\alias{test_for}
\alias{test_ifelse}
\alias{test_while}
\title{Test if student coded a control statement}
\usage{
test_ifelse(state, index = 1, not_found_msg = NULL)

test_while(state, index = 1, not_found_msg = NULL)

test_for(state, index = 1, not_found_msg = NULL)
}
\arguments{
\item{state}{state to start from}

\item{index}{Number of that particular control statement to check}

\item{not_found_msg}{Custom message in case the control statement was not found}
}
\description{
The \code{state} to start from is typically \code{ex()}, unless you're testing nested control statements.
}

