% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-control.R
\name{test_control_parts}
\alias{test_body.ControlState}
\alias{test_cond}
\alias{test_control_parts}
\alias{test_else}
\alias{test_if}
\title{Test elements of control statements}
\usage{
test_cond(state)

\method{test_body}{ControlState}(state)

test_if(state)

test_else(state, not_found_msg = NULL)
}
\arguments{
\item{state}{state to start from}

\item{not_found_msg}{custom message if the else part is not found}
}
\description{
Test elements of control statements
}

