% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-correct.R
\name{test_correct}
\alias{test_correct}
\title{Test things. If it fails, test additional things.}
\usage{
test_correct(check_code, diagnose_code)
}
\arguments{
\item{check_code}{Tests that are always run.}

\item{diagnose_code}{Tests that are run, only if the tests in \code{check_code} fail.}
}
\description{
Test if a set of tests passes, and do additional, more precise tests if there
were failures. The teacher should specify two code chunks; \itemize{ 
\item{\code{check_code}: specifies the code that checks on the (typically,
final results of the) student's code. These tests are executed silently,
without the reporter generating information for these.} 
\item{\code{diagnose_code}: Set of tests that gets executed if the tests in
\code{check_code} fail. These tests contain more detailed tests, to pinpoint
the problem.} }
}
\details{
\code{test_correct} increases the flexibility for the student: if the tests 
in \code{check_code} pass, the results of the tests in \code{diagnose_code}
are not considered. If you test for the end result in \code{check_code}, and
only do more rigorous testing in \code{diagnose_code}, you can allow for
multiple solutions to a challenge.
}
\examples{
\dontrun{
# Example solution code:
# x <- mean(1:3)

# Example SCT
test_correct({
 test_object("x")
}, {
 test_function("mean", "x")
})

# Following submissions will all be accepted:
x <- mean(1:3)
x <- 2
x <- mean(-1:5)
}

}

