% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-data-frame.R
\name{test_data_frame}
\alias{test_data_frame}
\title{Test whether a student correctly defined the column(s) of a data.frame}
\usage{
test_data_frame(name, columns, eq_condition = "equivalent",
  student_env = .GlobalEnv, solution_env = get_solution_env(),
  undefined_msg = NULL, undefined_cols_msg = NULL, incorrect_msg = NULL)
}
\arguments{
\item{name}{name of the data frame to test.}

\item{columns}{vector of names of columns to test}

\item{eq_condition}{character string indicating how to compare the
objects.  Possible values are \code{"equivalent"} (the default),
\code{"equal"} and \code{"identical"}.  See \code{\link{is_equivalent_to}},
\code{\link{equals}}, and \code{\link{is_identical_to}}, respectively.}

\item{student_env}{environment in which the student's code was evaluated.}

\item{solution_env}{environment in which the sample solution code was
evaluated.}

\item{undefined_msg}{feedback message in case the student did not define
the data frame.}

\item{undefined_cols_msg}{feedback message in case the student did define one or more fo the columns inside the data frame}

\item{incorrect_msg}{feedback message in case the columns queried does not correspond between student and solution

# Other examples: see SCT design guide}
}
\description{
Test whether a student defined a data.frame, and, if see, whether the columns of this data.frame
correspond to the sample solution.
}
\details{
This test is implemented using \code{\link{test_that}}.  Whether the
student's column and the sample solution object are the same is tested with
\code{\link{is_equivalent_to}}, hence small numeric differences or
differences in attributes are allowed.
}

