% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-equal.R
\name{test_equal}
\alias{test_equal}
\alias{test_equal.ArgumentState}
\alias{test_equal.ExprErrorState}
\alias{test_equal.ExprOutputState}
\alias{test_equal.ExprResultState}
\alias{test_equal.FunctionResultState}
\alias{test_equal.ObjectColumnState}
\alias{test_equal.ObjectElementState}
\alias{test_equal.ObjectState}
\alias{test_equal.OperationResultState}
\alias{test_equal.default}
\title{Test equality (general)}
\usage{
test_equal(state, ...)

\method{test_equal}{default}(state, ...)

\method{test_equal}{ObjectState}(state, incorrect_msg = NULL,
  eq_condition = "equivalent")

\method{test_equal}{ObjectColumnState}(state, incorrect_msg = NULL,
  eq_condition = "equivalent")

\method{test_equal}{ObjectElementState}(state, incorrect_msg = NULL,
  eq_condition = "equivalent")

\method{test_equal}{FunctionResultState}(state, eq_condition = "equivalent",
  incorrect_msg = NULL)

\method{test_equal}{OperationResultState}(state, eq_condition = "equivalent",
  incorrect_msg = NULL)

\method{test_equal}{ExprResultState}(state, incorrect_msg = NULL,
  eq_condition = "equivalent")

\method{test_equal}{ExprOutputState}(state, incorrect_msg = NULL)

\method{test_equal}{ExprErrorState}(state, incorrect_msg = NULL)

\method{test_equal}{ArgumentState}(state, incorrect_msg = NULL, eval = TRUE,
  eq_condition = "equivalent")
}
\arguments{
\item{state}{state to start from}

\item{...}{arguments passed to S3 implementations}

\item{incorrect_msg}{custom message in case the objects to do not correspond}

\item{eq_condition}{how to compare (\code{"equivalent"}, \code{"equal"} or \code{"identical"})}

\item{eval}{how to evaluate the arguments}
}
\description{
Test equality (general)
}

