% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_error}
\alias{test_error}
\title{Check whether the output of the student code contained an error.}
\usage{
test_error(contains_error = get_student_error(), incorrect_msg = NULL)
}
\arguments{
\item{contains_error}{boolean if the output contained an error.}

\item{incorrect_msg}{feeback message in case the output contained an error}
}
\description{
Check whether the output of the student code contained an error, based on the DM.contains.error variable.
This function is implemented using \code{\link{test_that}}.
}
\examples{
\dontrun{
# add in SCT:
test_error()
}
}

