% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-expression-error.R
\name{test_expression_error}
\alias{test_expression_error}
\title{Test error of expression}
\usage{
test_expression_error(expr, no_error_msg = NULL, incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression that is executed in both environments.}

\item{no_error_msg}{Optional feedback message when the expression does not
generate an error when run in the student's environment.}

\item{incorrect_msg}{Optional feedback message in case the evaluation is not the
same in both environments. Automatically generated if not specified.}
}
\description{
Test whether the given expression throws an error
}
\examples{
\dontrun{
# Example 1 solution code:
# my_fun <- function(a) { stopifnot(a > 0) }

# Test whether my_fun(-1) throws error
test_function_definition("my_fun", {
 test_expression_error(my_fun(-1))
})
}

}

