% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-expression-output.R
\name{test_expression_output}
\alias{test_expression_output}
\title{Test output of expression}
\usage{
test_expression_output(expr, student_env = .GlobalEnv,
  solution_env = get_solution_env(), incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression that is executed in both environments}

\item{student_env}{Environment in which the student's code was evaluated.}

\item{solution_env}{Environment in which the sample solution code was
evaluated.}

\item{incorrect_msg}{Feedback message in case the evaluation is not the
same in both environments. If none is given, a standard message will be
generated.}
}
\description{
This test tests whether the given expression gives the same output in
the student and the solution environment.
}

