% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-expression-result.R
\name{test_expression_result}
\alias{test_expression_result}
\title{Test results of expression}
\usage{
test_expression_result(expr, eq_condition = "equivalent",
  student_env = .GlobalEnv, solution_env = get_solution_env(),
  incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression that is executed in both environments}

\item{eq_condition}{Character vector indicating how to perform the
comparison of the results Possible values are \code{"equivalent"}
(the default), \code{"equal"} and \code{"identical"}.  See
\code{\link{is_equivalent_to}}, \code{\link{equals}}, and
\code{\link{is_identical_to}}, respectively.}

\item{student_env}{Environment in which the student's code was evaluated.}

\item{solution_env}{Environment in which the sample solution code was
evaluated.}

\item{incorrect_msg}{Feedback message in case the evaluation is not the
same in both environments. If none is given, a standard message will be
generated.}
}
\description{
This test tests whether the given expression gives the same result in
the student and the solution environment.
}

