% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-call-old.R
\name{test_function}
\alias{test_function}
\alias{test_function_v2}
\title{Test whether a student correctly called a function}
\usage{
test_function(name, args = NULL, index = 1, eval = TRUE,
  eq_condition = "equivalent", not_called_msg = NULL,
  args_not_specified_msg = NULL, incorrect_msg = NULL)

test_function_v2(name, args = NULL, index = 1, eval = TRUE,
  eq_condition = "equivalent", not_called_msg = NULL,
  args_not_specified_msg = NULL, incorrect_msg = NULL)
}
\arguments{
\item{name}{name of the function to test.}

\item{args}{character vector of argument names that the student should have
supplied in the function calls.}

\item{index}{integer that specifies which call of \code{name} in the solution 
code will be checked.}

\item{eval}{logical vector indicating whether the corresponding argument
should be evaluated before testing. Setting this to \code{FALSE} can be
useful, e.g., to test whether the student supplied a large predefined
object, as only the corresponding \code{\link{name}} is compared in this
case (use with care!).}

\item{eq_condition}{character vector indicating how to perform the
comparison for each argument. See \code{\link{is_equal}}}

\item{not_called_msg}{custom feedback message in case the student did not call the
function often enough.}

\item{args_not_specified_msg}{custom feedback message in case the student did call the function
with the arguments listed in \code{args}}

\item{incorrect_msg}{custom feedback message in case the student did not call the
function with the same argument values as in the sample solution. 
You can specify a vector of arguments with the same length as \code{args}, to have
argument-specific custom feedback.}
}
\description{
Test whether a student called a function, possibly with certain arguments, 
correctly. Note that \code{test_function} and \code{test_function_v2} are now
identical and either can be used, although the latter is likely to be phased
out in the future.
}
\examples{
\dontrun{
# Suppose the solution contains: mean(1:3, na.rm = TRUE)
# To test this submission, provide the following in the sct
test_function("mean", c("x", "na.rm"))
}

}

