% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-function-definition.R
\name{test_function_definition}
\alias{test_function_definition}
\title{Check whether the student defined a function correctly}
\usage{
test_function_definition(name, in_and_out, student_code, student_env,
  undefined_msg, incorrect_msg)
}
\description{
Check whether the student defined a function correctly
}

