% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-function-old.R
\name{test_function_old}
\alias{test_function_old}
\title{Test whether a student correctly called a function}
\usage{
test_function_old(name, args = NULL, ignore = NULL, allow_extra = TRUE,
  eval = TRUE, eq_condition = "equivalent", not_called_msg = NULL,
  incorrect_msg = NULL)
}
\arguments{
\item{name}{name of the function to test.}

\item{args}{character vector of argument names that the student should have
supplied in the function calls.  If no argument names are given, it is only
tested whether the student called the function at least as many times as in
the sample solution.}

\item{ignore}{character vector of argument names that should not be tested
(useful in combination with \code{allow_extra = FALSE} to allow certain
arguments to be ignored, but not others).}

\item{allow_extra}{indicates whether extra arguments not specified by
\code{args} or \code{ignore} are allowed in the student's function calls.}

\item{eval}{logical vector indicating whether the corresponding argument
should be evaluated before testing.  Setting this to \code{FALSE} can be
useful, e.g., to test whether the student supplied a large predefined
object, as only the corresponding \code{\link{name}} is compared in this
case (use with care!).}

\item{eq_condition}{character vector indicating how to perform the
comparison for each argument. See \code{\link{test_object}}}

\item{not_called_msg}{feedback message in case the student did not call the
function at least as often as in the solution code.}

\item{incorrect_msg}{feedback message in case the student did not call the
function with the same argument values as in the sample solution.  If
there are multiple function calls in the sample solution, a vector of
feedback messages can be supplied.}
}
\description{
Test whether a student called a function with certain arguments as least as
many times as in a sample solution.
}
\examples{
\dontrun{
# Suppose the solution contains: mean(1:3, na.rm = TRUE)
# To test this submission, provide the following in the sct
test_function("mean", c("x", "na.rm"))
}

}

