% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-function-v2.R
\name{test_function_v2}
\alias{test_function_v2}
\title{Test whether a student correctly called a function (v2)}
\usage{
test_function_v2(name, args = NULL, ignore = NULL, allow_extra = TRUE,
  eval = TRUE, index = NULL, eq_condition = "equivalent",
  not_called_msg = NULL, incorrect_msg = NULL,
  incorrect_number_of_calls_msg = NULL, index_not_called_msg = NULL)
}
\arguments{
\item{name}{name of the object to test.}

\item{args}{character vector of argument names that the student should have
supplied in the function calls.  If no argument names are given, it is only
tested whether the student called the function at least as many times as in
the sample solution.}

\item{ignore}{character vector of argument names that should not be tested
(useful in combination with \code{allow_extra = FALSE} to allow certain
arguments to be ignored, but not others).}

\item{allow_extra}{indicates whether extra arguments not specified by
\code{args} or \code{ignore} are allowed in the student's function calls.}

\item{eval}{Next to existence, check if the value of the object corresponds
between student en solution environment.}

\item{index}{Which command to check on (both in solution and student).
If index is specified, it is checked whether solution and student contain
the same amount of function calls.}

\item{eq_condition}{character string indicating how to compare. Possible values 
are \code{"equivalent"} (the default), \code{"equal"} and \code{"identical"}.
See \code{\link{expect_equivalent}}, \code{\link{expect_equal}}, 
and \code{\link{expect_identical}}, respectively.}

\item{not_called_msg}{feedback message in case the student did not call the
function at least as often as in the solution code.}

\item{incorrect_msg}{optional feedback message in case the student's object is not
the same as in the sample solution. Only used if \code{eval} is \code{TRUE}. 
A meaningful message is automatically generated if not supplied.}

\item{incorrect_number_of_calls_msg}{feedback message in case the number of commands
in the solution does not correspond to the solution. (only used if index is not NULL.)}

\item{index_not_called_msg}{feedback message in case the queried index wasn't found.}
}
\description{
Test whether a student called a function with certain arguments as least as
many times as in a sample solution.
}
\details{
Only difference between V1 and V2 is how \code{index} is treated.
}
\examples{
\dontrun{
# Suppose the solution contains: mean(1:3, na.rm = TRUE)
# To test this submission, provide the following in the sct
test_function("mean", c("x", "na.rm"))
}

}

