% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-instruction.R
\name{test_instruction}
\alias{test_instruction}
\title{Test a single instruction of the challenges interface}
\usage{
test_instruction(index, code, env = parent.frame())
}
\arguments{
\item{index}{the instruction index}

\item{code}{the test code for that instruction}

\item{env}{environment in which to execute tests.}
}
\description{
See vignette for more information.
}

