% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-library-function.R
\name{test_library_function}
\alias{test_library_function}
\title{Test whether the library function was called correctly}
\usage{
test_library_function(package, not_called_msg = NULL, incorrect_msg = NULL)
}
\arguments{
\item{package}{package name for which the library()
function should've been called}

\item{not_called_msg}{optional feedback message in case the library
function wasn't called a single time}

\item{incorrect_msg}{optional feedback message in case the library
function wasn't called for the specified package.}
}
\description{
Convenience function to test in a very hacky way whether 
the library function was called correctly in its most simple form.
There is support for the different ways to call the library function
}
\examples{
\dontrun{
# Example solution code
library(ggvis)
library(dplyr)

# SCT to test both library calls:
test_library_function("ggvis")
test_library_function("dplyr")
}

}
