% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-loop.R
\name{test_loop}
\alias{test_for_loop}
\alias{test_loop}
\alias{test_while_loop}
\title{Test a \code{while} or \code{for} loop}
\usage{
test_while_loop(index = 1, cond_test = NULL, expr_test = NULL,
  not_found_msg = NULL, env = parent.frame())

test_for_loop(index = 1, cond_test = NULL, expr_test = NULL,
  not_found_msg = NULL, env = parent.frame())
}
\arguments{
\item{index}{The index of the \code{while} or \code{for} loop to check (counting for for and while loops is separate)}

\item{cond_test}{SCT to perform on the condition part}

\item{expr_test}{SCT to perform on the expression part}

\item{not_found_msg}{Message in case the loop (at the specified index) is not found.}

\item{env}{Environment in which to perform all these SCTs}
}
\description{
Test whether a student correctly coded a while or \code{for} loop. The function parses
the student and solution code and selects a loop you specify based on the\code{index} argument, 
and then runs two chunks of tests:
\itemize{
 \item{\code{cond_test}: \code{testwhat} tests specifically for
 the condition part of the while or \code{for} loop.}
 \item{\code{expr_test}: \code{testwhat} tests for the body of the \code{for} or \code{while} loop}
}
The tests for the conditional part and the expression part
of the while loop can only be text-based. You cannot appropriately use functions such
as \code{\link{test_object}} that also depend on the student and solution
environment.
}
\examples{
\dontrun{
# while loop - example solution code:
while(x < 18) {
 x <- x + 5
 print(x)
}

# SCT to test this loop:
test_while_loop(1, cond_test = {
 test_student_typed(c("< 18", "18 >"))
}, expr_test = {
 test_student_typed(c("x + 5", "5 = x"))
 test_function("print", eval = FALSE) # no actual value matching possible!!
})

# for loop - example solution code:
for(i in 1:5) {
 print("hurray!") 
}

# SCT to test this loop:
test_for_loop(1, cond_test = {
 test_student_typed("in")
 test_student_typed("1")
 test_student_typed("5")
}, expr+test = {
 test_function("print")
})
}
}

