% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-object.R
\name{test_object}
\alias{test_object}
\title{Test whether a student correctly defined an object}
\usage{
test_object(name, eq_condition = "equivalent", student_env = .GlobalEnv,
  solution_env = get_solution_env(), undefined_msg = NULL,
  incorrect_msg = NULL)
}
\arguments{
\item{name}{name of the object to test.}

\item{eq_condition}{character string indicating how to compare the
objects.  Possible values are \code{"equivalent"} (the default),
\code{"equal"} and \code{"identical"}.  See \code{\link{is_equivalent_to}},
\code{\link{equals}}, and \code{\link{is_identical_to}}, respectively.}

\item{student_env}{environment in which the student's code was evaluated.}

\item{solution_env}{environment in which the sample solution code was
evaluated.}

\item{undefined_msg}{feedback message in case the student did not define
the object.}

\item{incorrect_msg}{feedback message in case the student's object is not
the same as in the sample solution.}
}
\description{
Test whether a student defined a certain object and, if yes, test whether
that object is the same as in a sample solution.
}
\details{
This test is implemented using \code{\link{test_that}}.  Whether the
student's object and the sample solution object are the same is tested with
\code{\link{is_equivalent_to}}, hence small numeric differences or
differences in attributes are allowed.

See vignette \code{"testwhat-intro"} for examples.
}
\examples{
\dontrun{
# Suppose the solution contains: x <- mean(1:3, na.rm = TRUE)
# To test this submission, provide the following in the sct
test_function("x")
}

# Other examples: see SCT design guide
}

