% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-or.R
\name{test_or}
\alias{test_or}
\title{Test if one of the given sct parts are correct.}
\usage{
test_or(..., incorrect_msg = NULL, choose_feedback = 1,
  env = parent.frame())
}
\arguments{
\item{...}{one of these code blocks with tests should succeed}

\item{incorrect_msg}{msg displayed when none succeeds}

\item{choose_feedback}{choose feedback of test with this index}

\item{env}{environment in which to execute tests.}
}
\description{
Test if one of the given SCT code batteries are evaluated as being
correct. If not, the feedback message of the first fail is standardly
given. Can be used nested.
}
\details{
\itemize{
 \item{\code{...}: an arbritrary amount of code blocks containing SCT 
 code. \code{test_or} will check if one of the code blocks results in 
 a successful SCT evaluation.}
}
}
\examples{
\dontrun{
  # test if either the object a or the object b is correct
  test_or(test_object("a"), test_object("b"))
}

}

