% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-output.R
\name{test_output.default}
\alias{test_output.default}
\title{Test output for regex}
\usage{
\method{test_output}{default}(state, regex, fixed = FALSE, trim = FALSE,
  times = 1, missing_msg = NULL)
}
\description{
Test output for regex
}

