% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-regex-old.R
\name{test_output_contains}
\alias{test_output_contains}
\title{Check whether the student printed something to the console}
\usage{
test_output_contains(expr, times = 1, incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression (as string) for which the output should be in the student's console output.}

\item{times}{How often the expression's output should occur in the student's console}

\item{incorrect_msg}{feeback message in case the output did not contain the expression}
}
\description{
Function checks whether the student's console contains the output one gets by
evaluating the character string provided in \code{expr} provided to expr. 
This function needs refactoring, as all new lines etc are removed.
}
\examples{
\dontrun{
# SCT to test whether student printed numbers 1 to 10
test_output_contains("for(i in 1:10) print(i)")
}

}

