% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-output-regex.R
\name{test_output_regex}
\alias{test_output_regex}
\title{Check whether the student output contains a pattern}
\usage{
test_output_regex(pattern, console_output = get_student_output(),
  incorrect_msg = NULL)
}
\arguments{
\item{pattern}{The pattern to check for in the console output.}

\item{console_output}{The string containing the output printed to the student's console.}

\item{incorrect_msg}{feeback message in case the pattern was not found in the console output}
}
\description{
Check whether the student output contains a pattern
}

