% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-test-pipe.R
\name{test_pipe}
\alias{test_pipe}
\title{Test whether a student used the pipe operator sufficiently (ggvis and dplyr exercises)}
\usage{
test_pipe(num = 1, absent_msg = NULL, insuf_msg = NULL)
}
\arguments{
\item{num}{minimal number of times the pipe operator has to appear (default = 1)}

\item{absent_msg}{feedback message in case the student did not use a single pipe.}

\item{insuf_msg}{feeback message in case the student did not use the pipe operator sufficiently.}
}
\description{
Test whether a student used the pipe sufficiently. By default, the function only checks if
the pipe was used at least once. The user can also select the minimal
number of occurrences of the pipe.
}
