% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-rmd-file.R
\name{test_rmd_file}
\alias{test_rmd_file}
\title{Test R Markdown file}
\usage{
test_rmd_file(code, student_file = NULL, solution_file = NULL,
  env = parent.frame())
}
\arguments{
\item{code}{the SCT code for the file}

\item{student_file}{the name of the student file to be tested}

\item{solution_file}{the name of the solution file to be tested}

\item{env}{The environment in which the code should be tested.}
}
\description{
Test a single R Markdown file
}
\details{
This test should be called when there are multiple files in the submission.
}

