% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-test-rmd-group.R
\name{test_rmd_group}
\alias{test_rmd_group}
\title{Test a single R Markdown file group (R Markdown exercises)}
\usage{
test_rmd_group(group_number, code)
}
\arguments{
\item{group_number}{Number of the group.}

\item{code}{SCT code to test the group (in curly braces)}
}
\description{
Test a single R Markdown file group (R Markdown exercises) with arbitrary testwhat functions.
}
\keyword{internal}
