% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-student-typed.R
\name{test_student_typed}
\alias{test_student_typed}
\title{Test whether a student typed something}
\usage{
test_student_typed(strings, student_code = get_student_code(),
  not_typed_msg = NULL)
}
\arguments{
\item{strings}{A set of strings, at least one of which must be in the student_code}

\item{student_code}{student submission as a string}

\item{not_typed_msg}{Feedback message in case the student did not type the string.}
}
\description{
Test whether a student typed something
}

