% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-test-subexpr-eval.R
\name{test_subexpr_eval}
\alias{test_subexpr_eval}
\title{Test whether a student called a subexpression correctly. (dplyr and ggvis exercises)}
\usage{
test_subexpr_eval(index = 1, fun = NULL, not_called_msg = NULL,
  incorrect_msg = NULL, incorrect_number_of_calls_msg = NULL)
}
\arguments{
\item{index}{exercise to be checked (solution and student code should have same number of calls!)}

\item{fun}{name of the function to be checked. if fun = NULL, check the entire command.}

\item{not_called_msg}{feedback message in case the function is not retrieved.}

\item{incorrect_msg}{feedback message in case the evaluation was not the same as in the solution}

\item{incorrect_number_of_calls_msg}{feedback message in case the student did
enter the same amount of commands as the solution did.}
}
\description{
Test whether a student called a(n) (sub)expression. If yes, test for this function call
if the result corresponds to the subexpression called in the solution.
}
\details{
This test is implemented using \code{\link{test_that}}.  When testing
whether the resut is the same, small numeric differences or
differences in attributes are allowed.
}

