% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-sufficient-length.R
\name{test_sufficient_length}
\alias{test_sufficient_length}
\title{Helper function to test if list has sufficient length}
\usage{
test_sufficient_length(stud, index, incorrect_number_of_calls_msg = NULL)
}
\arguments{
\item{stud}{list of student commands}

\item{index}{index to fetch from these commands}

\item{incorrect_number_of_calls_msg}{message in case the length of stud is not sufficient.}
}
\description{
Helper function to test if list has sufficient length
}

