% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-test-text.R
\name{test_text}
\alias{test_text}
\title{Test inline text and formatting (Markdown)}
\usage{
test_text(text, format = "any", freq = 1, not_called_msg = NULL,
  incorrect_msg = NULL)
}
\arguments{
\item{text}{Text to match (can be a regular expression!)}

\item{format}{the format of the text that the text should be in ("any", "italics", "bold", "code", "inline_code", "brackets", "parentheses", "list"). 
If none of the above, the format string is appended to text in front and in the back and used as a regexp.}

\item{freq}{How often the text should appear with this formatting}

\item{not_called_msg}{feedback message if the text was not there}

\item{incorrect_msg}{feedback message if the text was not properly formatted}
}
\description{
Test inline text and its formatting for R Markdown exercises.
This test can only be called inside a test_rmd_group() call!
}
\keyword{internal}
