% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-that.R
\name{test_that}
\alias{test_that}
\title{Overridden version of testthat's function for compatibility reasons}
\usage{
test_that(descr, code)
}
\arguments{
\item{descr}{description}

\item{code}{actual test code}
}
\description{
Overridden version of testthat's function for compatibility reasons
}

