% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-test-tree-contains.R
\name{test_tree_contains}
\alias{test_tree_contains}
\title{Test whether a student's function call contains a certain character (dplyr and ggvis exercises)}
\usage{
test_tree_contains(index = 1, fun = NULL, queries = NULL, times = NULL,
  contain_all = TRUE, fixed_order = FALSE, not_called_msg = NULL,
  absent_msg = NULL, incorrect_number_of_calls_msg = NULL)
}
\arguments{
\item{index}{exercise to be checked (solution and student code should have same number of calls!)}

\item{fun}{name of the function to be checked. if fun = NULL, check the entire command.}

\item{queries}{single character or vector of character that have to be present in the function}

\item{times}{number of times each of the entered queries have to be available. (ones on default)}

\item{contain_all}{AND versus OR. if contain_all = TRUE (by default), all strings in the queries
vector must exist in the tree. If contain_all = FALSE, the test passes if one of the strings in the
queries vector exists in the tree the specified number of times.}

\item{fixed_order}{whether or not the queries have to appear in the function call the same order
as specified in the queries vector (default FALSE).
This functionality can not be used when contain_all is FALSE.
This option also only works when each query has to be present only once.
The order is determined on the LAST occurrences of the queries (in embedded notation!!)}

\item{not_called_msg}{feedback message in case the function is not retrieved}

\item{absent_msg}{feedback message in case one of the queries was not available}

\item{incorrect_number_of_calls_msg}{feedback message in case the student did
enter the same amount of commands as the solution did.}
}
\description{
For a specified command in the student's code, check whether a particular function contains
a set of queries. Also the number of times these queries have to appear can be specified.
}
\details{
Only exact string matching is performed for the moment.
}

