% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-test-tree-order.R
\name{test_tree_order}
\alias{test_tree_order}
\title{Test whether a student's collection of function calls follows that of the solution (dplyr exercises)}
\usage{
test_tree_order(index = 1, custom_orders = NULL, allow_extra = TRUE,
  incorrect_msg = NULL, incorrect_number_of_calls_msg = NULL)
}
\arguments{
\item{index}{exercise to be checked (solution and student code should have same number of calls!)}

\item{custom_orders}{list of character vectors. Every vector represents an order that is acceptable.
Example of syntax to define orders: list(c("select","filter","arrange"),c("filter","arrange","select")).
This specific example would mean that a function call where a select inside filter inside arrange,
or a filter inside arrange or inside select is allowed.}

\item{allow_extra}{if TRUE, allows the student to define additional function, as long as the functions in
the solution are followed in the same order.}

\item{incorrect_msg}{feedback message in case the student order did not match any of the accepted orders.}

\item{incorrect_number_of_calls_msg}{feedback message in case the student did
enter the same amount of commands as the solution did.}
}
\description{
For a specified command in the student's code, check whether the order of function calls
follows the one specified by the teacher. The teacher can set several orders that are accepted.
Only exact string matching is performed for the moment. No support for summarize(), only summarise()! 
(unless teacher codes it explicitly)
}

