% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-what.R
\name{test_what}
\alias{test_what}
\title{Expectation wrapper}
\usage{
test_what(code, feedback_msg)
}
\arguments{
\item{code}{The expectation that should be wrapped}

\item{feedback_msg}{The feedback message that results when the expectation is not met}
}
\description{
This function wraps around an expect_... function. When the expectation fails to be
met, the feedback message is sent to the reporter.
}

