% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-loop.R
\name{test_while_loop}
\alias{test_while_loop}
\title{Check whether the student wrote a while loop correctly}
\usage{
test_while_loop(var, seq, sct_expr, undefined_msg, incorrect_msg)
}
\description{
Check whether the student wrote a while loop correctly
}

