% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{tw}
\alias{tw}
\title{tw singleton object to access data across SCT chains.}
\format{An object of class \code{list} of length 4.}
\usage{
tw
}
\description{
tw singleton object to access data across SCT chains.
}
\keyword{datasets}
