% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rmd.R
\name{build_doc_structure}
\alias{build_doc_structure}
\title{build R markdown document structure, using knitr functions}
\usage{
build_doc_structure(text)
}
\arguments{
\item{text}{text representing an R Markdown document}
}
\description{
build R markdown document structure, using knitr functions
}

