% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pd.R
\name{build_pd}
\alias{build_pd}
\title{Build ParseData from string representing code}
\usage{
build_pd(code, silent = TRUE)
}
\arguments{
\item{code}{character string representing code}

\item{silent}{if TRUE, parsing errors are caugt and NULL is returned.}
}
\description{
Build ParseData from string representing code
}
