% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-function-result.R
\name{check_function_result}
\alias{check_function_result}
\alias{check_result.OperationState}
\alias{check_result.FunctionState}
\alias{check_equal.FunctionResultState}
\alias{check_equal.OperationResultState}
\title{Check the result of a function call/operation}
\usage{
\method{check_result}{OperationState}(state, error_msg = NULL,
  append = TRUE, ...)

\method{check_result}{FunctionState}(state, error_msg = NULL, append = TRUE,
  ...)

\method{check_equal}{FunctionResultState}(state, eq_condition = "equivalent",
  incorrect_msg = NULL, append = TRUE, ...)

\method{check_equal}{OperationResultState}(state, eq_condition = "equivalent",
  incorrect_msg = NULL, append = TRUE, ...)
}
\arguments{
\item{state}{the state to start from (for \code{check_} functions)}

\item{error_msg}{feedback message in case the student function call at the mentioned index generated an error.}

\item{append}{Whether or not to append the feedback to feedback built in previous states}

\item{...}{S3 stuff}

\item{eq_condition}{how to compare student and solution.}

\item{incorrect_msg}{feedback message in case the evaluation was not the same as in the solution.}
}
\description{
Check the result of a function call/operation
}
\examples{
\dontrun{
# Example 1
mean(1:3)

# SCT
ex() \%>\% check_function("mean") \%>\% check_result() \%>\% check_equal()
}
}
