% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.r
\name{get_clean_lines}
\alias{get_clean_lines}
\title{convert student/solution code to vector of clean strings with the pipe operator removed.}
\usage{
get_clean_lines(code)
}
\description{
convert student/solution code to vector of clean strings with the pipe operator removed.
}

