% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expectations-testwhat.R
\name{is_defined}
\alias{expect_defined}
\alias{is_defined}
\title{Expectation: is an object defined?}
\usage{
is_defined(env = .GlobalEnv)

expect_defined(object, env = .GlobalEnv, info = NULL, label = NULL)
}
\arguments{
\item{env}{environment in which to look for the object.}

\item{object}{name of the object to test}

\item{label}{object label used in the default feedback message.  If
\code{NULL}, the object name is used.  This is ignored for feedback
messages supplied via \code{failure_msg} or \code{success_msg}.}
}
\description{
Tests whether or not an object is defined in a certain environment.
}
\seealso{
\code{\link{exists}}
}

