% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_docs}
\alias{parse_docs}
\title{Parse both the student and solution document}
\usage{
parse_docs(student_code = get_student_code(),
  solution_code = get_solution_code())
}
\description{
Parse both the student and solution document
}

