% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-exercise.R
\name{run_until_fail}
\alias{run_until_fail}
\title{Run SCT until it fails}
\usage{
run_until_fail(code)
}
\arguments{
\item{code}{the SCT script to run as an expression}
}
\description{
Run SCT until it fails
}
