% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-s3-definitions.R
\name{s3defs}
\alias{s3defs}
\alias{check_equal}
\alias{check_equal.default}
\alias{check_result}
\alias{check_result.default}
\alias{check_body}
\alias{check_body.default}
\alias{check_option}
\alias{check_option.default}
\title{S3 definitions}
\usage{
check_equal(state, ...)

\method{check_equal}{default}(state, ...)

check_result(state, ...)

\method{check_result}{default}(state, ...)

check_body(state, ...)

\method{check_body}{default}(state, ...)

check_option(state, ...)

\method{check_option}{default}(state, ...)
}
\arguments{
\item{state}{State to start from}

\item{...}{Additional arguments passed to implementations of the S3 method}
}
\description{
S3 definitions
}
