% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages-utils.R
\name{set_language}
\alias{set_language}
\title{Specify the language for the automated feedback}
\usage{
set_language(lang)
}
\arguments{
\item{lang}{shorthand notation for the language. Currently, "en", "fr" and "es" are supported}
}
\description{
Specify the language for the automated feedback
}

