% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-state.R
\name{setup_state}
\alias{setup_state}
\title{Set up state for local experimentation.}
\usage{
setup_state(sol_code, stu_code, pec = character(),
  ex_type = "NormalExercise")
}
\arguments{
\item{sol_code}{Solution script as a string}

\item{stu_code}{Student submission as a string. If it is not specified, the
solution code will be used.}

\item{pec}{Pre-exercise-code as a string}

\item{ex_type}{Type of exercise as a string. Defaults to NormalExercise.}
}
\value{
The exercise state, from which you can start chaining.
}
\description{
It runs both the solution and the student submission, and populates the state
with parse data, output, etc. After running this function, the state is
available thorugh \code{\link{ex}}, from which you can start your SCT chains.
In a way, this function is a very light weight version of DataCamp's R Backend.
}
\note{
This function is only supposed to be used locally when experimenting.
  It should never be used in the eventual SCT script of an exercise.
}
\examples{
\dontrun{
setup_state(
 sol_code = "a <- 1",
 stu_code = "a <- 2"
)

ex() \%>\% check_object('a') \%>\% check_equal()
}

}
