% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/success-msg.R
\name{success_msg}
\alias{success_msg}
\title{Define the success message}
\usage{
success_msg(msg)
}
\arguments{
\item{msg}{The congratulatory message as a character string.}
}
\description{
If all tests in an SCT pass, the students gets a
congratulatory message. You can specify this message with
/code{success_msg()}. It does not matter where in the SCT
you specify this message, but at the end makes most sense.
}
\details{
For multiple choice exercises, the success message is specified
inside \code{\link{test_mc}}, so an additional call of \code{success_msg}
is not necessary.
}

