% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-an-object.R
\name{test_an_object}
\alias{test_an_object}
\title{Check if the student defined an object, independent of the name}
\usage{
test_an_object(name, undefined_msg = NULL, eq_condition = "equivalent",
  student_env = .GlobalEnv, solution_env = get_solution_env())
}
\arguments{
\item{name}{name of object in solution to test.}

\item{undefined_msg}{feedback message in case the student did not define 
an object that corresponds to the solution object. This argument should 
always be specified.}

\item{eq_condition}{character string indicating how to compare. Possible values 
are \code{"equivalent"} (the default), \code{"equal"} and \code{"identical"}.
See \code{\link{expect_equivalent}}, \code{\link{expect_equal}}, 
and \code{\link{expect_identical}}, respectively.}

\item{student_env}{Environment in which the student code was evaluated. Do not set this manually.}

\item{solution_env}{Environment in which the solution code was evaluated. Do not set this manually.}
}
\description{
This function is an adaption of \code{\link{test_object}}. The function will
check if a specific object, defined in the solution, exists. The object the 
student defined doesn't have to have the same name. In other words,
this function will check if any defined variable by the user corresponds to
a specific variable in the solution code.
}
\examples{
\dontrun{
# Example 1 solution code:
# x <- 5

# sct command to test whether student defined _an_ object with same value
test_an_object("x")

# All of the following student submissions are accepted
# x <- 5
# y <- 5
# z <- 4 + 1 + 1e-8
}

}

