% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-data-frame.R
\name{test_data_frame}
\alias{test_data_frame}
\title{Test list elements (or data frame columns)}
\usage{
test_data_frame(name, columns = NULL, eq_condition = "equivalent",
  student_env = .GlobalEnv, solution_env = get_solution_env(),
  undefined_msg = NULL, undefined_cols_msg = NULL, incorrect_msg = NULL)
}
\arguments{
\item{name}{name of the list or data frame to test.}

\item{columns}{character vector or integer vector of list elements or
indices to test.}

\item{eq_condition}{character string indicating how to compare. Possible values
are \code{"equivalent"} (the default), \code{"equal"} and \code{"identical"}.
See \code{\link{expect_equivalent}}, \code{\link{expect_equals}},
and \code{\link{expect_identical}}, respectively.}

\item{student_env}{environment in which the student code was evaluated.}

\item{solution_env}{environment in which the solution code was evaluated.}

\item{undefined_msg}{optional feedback message if list is not defined.}

\item{undefined_cols_msg}{optional feedback message if not all specified elements
of the solution list were found in the student's list.}

\item{incorrect_msg}{optional feedback message if not all specified elements of
the solution list match those in the student list.}
}
\description{
Test whether a student defined a list, and if this is the case,
whether the elements of the list correspond to the ones in the solution.
A data frame is also a list, so you can use this function to test the
correspondence of data frame columns.
}
\examples{
\dontrun{
# Example 1 solution code:
# df <- data.frame(a = 1:3, b = LETTERS[1:3])

# sct command to test column a
test_data_frame("df", columns = "a")

# sct command to test column b
test_data_frame("df", columns = "b")
}
}

