% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-error.R
\name{test_error}
\alias{check_error.default}
\alias{test_error}
\title{Check whether the student's submission threw an error.}
\usage{
test_error()

\method{check_error}{default}(state, ...)
}
\arguments{
\item{state}{State to start from (for \code{check_error})}

\item{...}{S3 stuff}
}
\description{
With information gathered from the R Backend, detect
whether the student's submission generated an error.
}
\examples{
\dontrun{
# Example student code: x <- 4 + "a"

# SCT option 1
test_error()

# SCT option 2
ex() \%>\% check_error()
}

}

