% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-error.R
\name{test_error}
\alias{test_error}
\alias{test_error.ExprState}
\alias{test_error.FunDefState}
\alias{test_error.default}
\title{Check whether the student's submission threw an error.}
\usage{
test_error(state, ...)

\method{test_error}{default}(state)

\method{test_error}{ExprState}(state, no_error_msg = NULL)

\method{test_error}{FunDefState}(state, ..., no_error_msg = NULL)
}
\arguments{
\item{state}{state to start from (can be missing)}

\item{...}{arguments that are used to build a function definition call}

\item{no_error_msg}{custom message in case the expression or function call didn't generate an error}
}
\description{
With information gathered from the R Backend, \code{test_error} detects
whether the student's submission generated an error. Automatically, a feedback is generated,
which can be appended with an additional incorrect_msg.
}
\examples{
\dontrun{
# Example student code: x <- 4 + "a"

# R error message as feedback:
test_error()
}

}

