% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-error.R
\name{test_error}
\alias{test_error}
\title{Check whether the student's submission threw an error.}
\usage{
test_error(incorrect_msg = NULL, error = get_student_error())
}
\arguments{
\item{incorrect_msg}{feeback message that is appended to the error message generated by R.}

\item{error}{The error message if the student's submission generated one;
NULL if there was no error. Do not set this manually.}
}
\description{
With information gathered from the R Backend, \code{test_error} detects
whether the student's submission generated an error. Automatically, a feedback is generated,
which can be appended with an additional incorrect_msg.
}
\examples{
\dontrun{
# Example student code: x <- 4 + "a"

# R error message as feedback:
test_error()

# R error message as feedback, with additional info:
test_error("Don't sum numerics and characters!")
}
}

