% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-error.R
\name{test_error}
\alias{test_error}
\title{Check whether the student's submission threw an error.}
\usage{
test_error(state, ...)
}
\description{
With information gathered from the R Backend, \code{test_error} detects
whether the student's submission generated an error. Automatically, a feedback is generated,
which can be appended with an additional incorrect_msg.
}
\examples{
\dontrun{
# Example student code: x <- 4 + "a"

# R error message as feedback:
test_error()
}

}

