% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-exercise.R
\name{test_exercise}
\alias{failure_msg}
\alias{success_msg}
\alias{test_exercise}
\title{Run all tests for an exercise}
\usage{
test_exercise(code, report = c("first", "all", "challenge"),
  failure_msg = NULL, success_msg = NULL, env = test_env())

failure_msg(msg)

success_msg(msg)
}
\arguments{
\item{code}{character string containing the tests to perform.}

\item{report}{specifies how to report feedback in case of failed tests.  If
\code{"first"}, only the feedback message from the first failed test is
reported.  If \code{"all"}, feedback messages from all failed tests are
collected.}

\item{failure_msg}{if \code{report = "all"}, text to be included before the
collected feedback messages in case of failed tests.  Otherwise ignored.}

\item{success_msg}{feedback message in case all tests are successful.}

\item{env}{environment in which to execute tests.}

\item{msg}{see \dQuote{Details}.}
}
\value{
A list with components \code{passed} that indicates whether all
tests were sucessful, and \code{feedback} that contains a feedback message.
}
\description{
Run all tests for an exercise and report the results (including feedback).
}
\details{
Function \code{test_exercise()} allows to supply feedback messages with
arguments \code{failure_msg} (only if \code{report = "all"}) and
\code{success_msg}.  Alternatively, functions \code{failure_msg()} and
\code{success_msg()} can be called within the test code to set the
corresponding feedback messages.

See vignette \code{"testwhat-intro"} for examples.
}

