% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-exercise.R
\name{test_exercise}
\alias{test_exercise}
\title{Run all tests for an exercise}
\usage{
test_exercise(code, report = c("first", "challenge"), pec, student_code,
  solution_code, solution_env, output_list, env = test_env())
}
\arguments{
\item{code}{character string containing the tests to perform.}

\item{report}{specifies how to report feedback in case of failed tests.  If
\code{"first"}, only the feedback message from the first failed test is
reported. If \code{"all"}, feedback messages from all failed tests are
collected.}

\item{env}{environment in which to execute tests.}
}
\value{
A list with components \code{passed} that indicates whether all
tests were sucessful, and \code{feedback} that contains a feedback message.
}
\description{
Run all tests for an exercise and report the results (including feedback).
This function is run by R Backend and should not be used by course creators.
}

