% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-exercise.R
\name{test_exercise}
\alias{test_exercise}
\title{Run all tests for an exercise}
\usage{
test_exercise(sct, ex_type, pec, student_code, solution_code, student_env,
  solution_env, output_list, seed = 42)
}
\arguments{
\item{sct}{Submission correctness tests as a character string.}

\item{ex_type}{Type of the exercise}

\item{pec}{pre-exercise-code}

\item{student_code}{character string representing the student code}

\item{solution_code}{character string representing the solution code}

\item{student_env}{environment containing the objects defined by the student.}

\item{solution_env}{environment containing the objects defined by solution code}

\item{output_list}{the output structure that is generated by RBackend}

\item{seed}{random seed that is used for SCTs that run expressions
If this is the case, e.g. \code{\link{test_or}} will behave differently.}
}
\value{
A list with components \code{passed} that indicates whether all
tests were sucessful, and \code{feedback} that contains a feedback message.
}
\description{
Run all tests for an exercise and report the results (including feedback).
This function is run by R Backend and should not be used by course creators.
}
