% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-expr.R
\name{test_expression_result}
\alias{test_expression_result}
\alias{test_expression_output}
\alias{test_expression_error}
\alias{check_expr}
\alias{check_result.ExprState}
\alias{check_output.ExprState}
\alias{check_error.ExprState}
\alias{check_equal.ExprResultState}
\alias{check_equal.ExprOutputState}
\alias{check_equal.ExprErrorState}
\title{Check the result, output or errors thrown by an expression}
\usage{
test_expression_result(expr, eq_condition = "equivalent",
  incorrect_msg = NULL)

test_expression_output(expr, incorrect_msg = NULL)

test_expression_error(expr, no_error_msg = NULL, incorrect_msg = NULL)

check_expr(state, expr)

\method{check_result}{ExprState}(state, error_msg = NULL, append = TRUE,
  ...)

\method{check_output}{ExprState}(state, error_msg = NULL, append = TRUE,
  ...)

\method{check_error}{ExprState}(state, no_error_msg = NULL, append = TRUE,
  ...)

\method{check_equal}{ExprResultState}(state, incorrect_msg = NULL,
  append = TRUE, eq_condition = "equivalent", ...)

\method{check_equal}{ExprOutputState}(state, incorrect_msg = NULL,
  append = TRUE, ...)

\method{check_equal}{ExprErrorState}(state, incorrect_msg = NULL,
  append = TRUE, ...)
}
\arguments{
\item{expr}{the expression to run}

\item{eq_condition}{character string indicating how to compare. See
\code{\link{is_equal}}.}

\item{incorrect_msg}{custom message in case the result, output or error of
the expression does not correspond with the solution}

\item{no_error_msg}{custom message in case the expression doesn't throw an
error while it should}

\item{state}{state to start from (only for \code{check_} functions)}

\item{error_msg}{custom message in case the expression throws an error while
it shouldn't}

\item{append}{Whether or not to append the feedback to feedback built in previous states}

\item{...}{S3 stuff}
}
\description{
Run an expression in student and solution environment and compare the result,
output or error that is thrown by it.
}
\examples{
\dontrun{
# Example 1
a <- c(1, 2, 3, 4, 5, 6) 

# SCT option 1 to test if second and fourth element is ok:
test_expression_result("a[c(2, 4)]")

# SCT option 2
ex() \%>\% check_expr("a[c(2, 4)]") \%>\% check_result() \%>\% check_equal()

# Example 2
my_fun <- function() { print('hello') }

# SCT option 1 to test if my_fun() produces correct output:
test_expression_output("my_fun()")

# SCT option 2
ex() \%>\% check_expr("my_fun()") \%>\% check_output() \%>\% check_equal()
}
}
