% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-expr-old.R
\name{test_expression_output}
\alias{test_expression_output}
\title{Test output of expression}
\usage{
test_expression_output(expr, incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression that is executed in both environments.}

\item{incorrect_msg}{Optional feedback message in case the evaluation is not the
same in both environments. Automatically generated if not specified.}
}
\description{
Test whether the given expression gives the same output in 
the student and the solution environment.
}
\examples{
\dontrun{
# Example 1 solution code:
# my_fun <- function(a, b) { a + b }

# Test whether my_fun(1,2) and my_fun(1,2)
# give same _output_
test_function_definition("my_fun", {
 test_expression_output(my_fun(1,2))
 test_expression_output(my_fun(-1,-2))
})
}

}

