% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-expression-result.R
\name{test_expression_result}
\alias{test_expression_result}
\title{Test result of expression}
\usage{
test_expression_result(expr, eq_condition = "equivalent",
  incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression that is executed in both environments.}

\item{eq_condition}{character string indicating how to compare. See
\code{\link{is_equal}}.}

\item{incorrect_msg}{Optional feedback message in case the evaluation is not the
same in both environments. Automatically generated if not specified.}
}
\description{
Test whether the given expression gives the same result in 
the student and the solution environment.
}
\examples{
\dontrun{
# Example 1 solution code:
# my_fun <- function(a, b) { a + b }

# Test whether my_fun(1,2) and my_fun(1,2)
# give same _result_
test_function_definition("my_fun", {
 test_expression_result(my_fun(1,2))
 test_expression_result(my_fun(-1,-2))
})
}

}

