% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-expression-result.R
\name{test_expression_result}
\alias{test_expression_result}
\title{Test result of expression}
\usage{
test_expression_result(expr, eq_condition = "equivalent",
  student_env = .GlobalEnv, solution_env = get_solution_env(),
  incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression that is executed in both environments.}

\item{eq_condition}{character string indicating how to compare. Possible values
are \code{"equivalent"} (the default), \code{"equal"} and \code{"identical"}.
See \code{\link{expect_equivalent}}, \code{\link{expect_equals}},
and \code{\link{expect_identical}}, respectively.}

\item{student_env}{environment in which the student code was evaluated.}

\item{solution_env}{environment in which the solution code was evaluated.}

\item{incorrect_msg}{Optional feedback message in case the evaluation is not the
same in both environments. Automatically generated if not specified.}
}
\description{
Test whether the given expression gives the same result in
the student and the solution environment.
}
\examples{
\dontrun{
# Example 1 solution code:
# my_fun <- function(a, b) { a + b }

# Test whether my_fun(1,2) and my_fun(1,2)
# give same _result_
test_function_defintion({
 test_expression_result(my_fun(1,2))
 test_expression_result(my_fun(-1,-2))
})
}
}

