% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-loop.R
\name{test_for_loop}
\alias{test_for_loop}
\title{Check whether the student wrote a for loop correctly}
\usage{
test_for_loop(var, seq, sct_expr, undefined_msg, incorrect_msg)
}
\description{
Check whether the student wrote a for loop correctly
}

