% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-for-loop.R
\name{test_for_loop}
\alias{test_for_loop}
\title{Check whether the student wrote a for loop correctly}
\usage{
test_for_loop(index = 1, cond_test = NULL, expr_test = NULL,
  student_code = get_student_code(), solution_code = get_solution_code(),
  not_found_msg = NULL, env = parent.frame())
}
\description{
Check whether the student wrote a for loop correctly
}

