% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-function-definition.R
\name{test_function_definition}
\alias{test_function_definition}
\title{Check whether the student defined a function correctly}
\usage{
test_function_definition(name, function_test = NULL, body_test = NULL,
  student_env = .GlobalEnv, solution_env = get_solution_env(),
  student_code = get_student_code(), solution_code = get_solution_code(),
  undefined_msg = NULL, incorrect_number_arguments_msg = NULL,
  env = parent.frame())
}
\arguments{
\item{name}{The name of the function to test}

\item{function_test}{Set of testthat (testwhat) tests to perform on the function}

\item{body_test}{fill SCT for the function body}

\item{student_env}{environment in which the student's code was evaluated.}

\item{solution_env}{environment in which the sample solution code was
evaluated.}

\item{student_code}{character string containing the student's code.}

\item{solution_code}{character string containing the sample solution code.}

\item{undefined_msg}{Feedback message in case the specified function was not defined}

\item{incorrect_number_arguments_msg}{Feedback message in case the function does not have the correct number of arguments.}

\item{env}{Environment in which to perform the tests}
}
\description{
Check whether the student defined a function correctly
}

