% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-function-definition.R
\name{test_function_definition}
\alias{test_function_definition}
\title{Check whether the student defined a function correctly}
\usage{
test_function_definition(name, function_test = NULL, body_test = NULL,
  arguments = c(), check_defaults = TRUE, student_env = .GlobalEnv,
  solution_env = get_solution_env(), student_code = get_student_code(),
  solution_code = get_solution_code(), undefined_msg = NULL,
  incorrect_number_arguments_msg = NULL, incorrect_arguments_msg = NULL,
  env = parent.frame())
}
\description{
Check whether the student defined a function correctly
}

