% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test_if_else.R
\name{test_if_else}
\alias{test_if_else}
\title{Check whether the student wrote a conditional construct correctly}
\usage{
test_if_else(index = 1, if_cond_test = NULL, if_expr_test = NULL,
  else_expr_test = NULL, student_code = get_student_code(),
  solution_code = get_solution_code(), not_found_msg = NULL,
  missing_else_msg = NULL, env = parent.frame())
}
\description{
Check whether the student wrote a conditional construct correctly
}

