% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-object.R
\name{test_object}
\alias{test_object}
\alias{check_object}
\alias{check_column}
\alias{check_element}
\alias{check_equal.ObjectState}
\alias{check_equal.ObjectColumnState}
\alias{check_equal.ObjectElementState}
\title{Check R object existence and value}
\usage{
check_object(state, name, undefined_msg = NULL, append = TRUE)

check_column(state, col, col_missing_msg = NULL, append = TRUE)

check_element(state, el, el_missing_msg = NULL, append = TRUE)

\method{check_equal}{ObjectState}(state, incorrect_msg = NULL,
  append = TRUE, eq_condition = "equivalent", ...)

\method{check_equal}{ObjectColumnState}(state, incorrect_msg = NULL,
  append = TRUE, eq_condition = "equivalent", ...)

\method{check_equal}{ObjectElementState}(state, incorrect_msg = NULL,
  append = TRUE, eq_condition = "equivalent", ...)
}
\arguments{
\item{state}{the state to start from}

\item{name}{name of the object to test.}

\item{undefined_msg}{Optional feedback message in case the student did not 
define the object. A meaningful message is automatically generated if not 
supplied.}

\item{append}{Whether or not to append the feedback to feedback built in previous states}

\item{col}{name of column to check}

\item{col_missing_msg}{Custom message in case data frame column is missing}

\item{el}{name of element to check}

\item{el_missing_msg}{Custom message in case element is messing.}

\item{incorrect_msg}{Custom feedback message in case the student's object is
not the same as in the sample solution.}

\item{eq_condition}{character string indicating how to compare. See 
\code{\link{is_equal}}.}

\item{...}{S3 stuff}
}
\description{
Check whether a student defined a certain object (correctly)
}
\examples{
\dontrun{
# Example 1
x <- mean(1:3, na.rm = TRUE)

# SCT
ex() \%>\% check_object("x") \%>\% check_equal()

# Example 2
x <- mean(1:3, na.rm = TRUE)

# SCT option
ex() \%>\% check_object("x")

# Example 3
df <- data.frame(a = 1:3, b = LETTERS[1:3])

# SCT to test column a
ex() \%>\% check_object("df") \%>\% check_column("a") \%>\% check_equal()

# Example 4
lst <- list(a = 1, b = 2)

# SCT to test only element b
ex() \%>\% check_object("lst") \%>\% check_element("b") \%>\% check_equal()
}
}
