% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-object-old.R, R/test-object.R
\name{test_object}
\alias{test_obj}
\alias{test_object}
\title{Test R object existence and value}
\usage{
test_object(name, eq_condition = "equivalent", eval = TRUE,
  undefined_msg = NULL, incorrect_msg = NULL)

test_obj(state, name, undefined_msg = NULL)
}
\arguments{
\item{name}{name of the object to test.}

\item{eq_condition}{character string indicating how to compare. See
\code{\link{is_equal}}.}

\item{eval}{Next to existence, check if the value of the object corresponds 
between student en solution environment.}

\item{undefined_msg}{Optional feedback message in case the student did not 
define the object. A meaningful message is automatically generated if not 
supplied.}

\item{incorrect_msg}{optional feedback message in case the student's object 
is not the same as in the sample solution. Only used if \code{eval} is 
\code{TRUE}. A meaningful message is automatically generated if not 
supplied.}
}
\description{
Test whether a student defined a certain object. If this is the case, and if 
\code{eval} is \code{TRUE}, also check whether the value of the object 
matches that of the solution.

Test R object existence
}
\examples{
\dontrun{
# Example 1 solution code:
# x <- mean(1:3, na.rm = TRUE)

# sct command to test existence and value of x:
test_object("x")

# sct command to test only existence of x:
test_object("x", eval = FALSE)

# Example 2 solution code:
# y <- list(a = 2, b = 3, c = 4)

# Small numerical difference allowed + no check on attributes
test_object(y)

# Small numerical difference allowed + check attributes
test_object(y, eq_condition = "equals")

# No numerical difference allowed + check attributes
test_object(y, eq_condtion = "identical")
}

}

