% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-operator.R
\name{test_operator}
\alias{test_operator}
\title{Test whether a student correctly used an inline operator}
\usage{
test_operator(name, index = 1, eval = TRUE, eq_condition = "equivalent",
  not_called_msg = NULL, not_specified_msg = NULL, incorrect_msg = NULL)
}
\description{
Test whether a student correctly used an inline operator
}
\examples{
\dontrun{
# Todo add examples
}

}

