% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-call-old.R
\name{test_operator}
\alias{test_operator}
\title{Test whether a student correctly used an inline operator}
\usage{
test_operator(name, index = 1, eval = TRUE, eq_condition = "equivalent",
  not_called_msg = NULL, error_msg = NULL, incorrect_msg = NULL)
}
\arguments{
\item{name}{Name of the operator as a string, e.g. \code{"+"}}

\item{index}{integer that specifies which \code{name} operator that will be
checked in the solution}

\item{eval}{whether or not to evaluate the expression that is formed by the
operator}

\item{eq_condition}{character vector indicating how to perform the
comparison for each argument. See \code{\link{is_equal}}}

\item{not_called_msg}{custom feedback message in case the student did not
call the operator often enough.}

\item{error_msg}{custom feedback message in case the operation by the student caused an error.}

\item{incorrect_msg}{custom feedback message in case the student's evaluation
of the expression that is formed by the operator does not correspond with
the corresponding call in the solution}
}
\description{
Test whether a student correctly used an inline operator
}
\examples{
\dontrun{
# Suppose the solution contains: 4 + 5

# To test this submission, provide the following in the sct
test_operator("+")
}

}

