% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-correct.R
\name{test_or}
\alias{test_or}
\title{Test if one of many test sets pass.}
\usage{
test_or(..., incorrect_msg = NULL, choose_feedback = 1)
}
\arguments{
\item{...}{a set of tests, each set separated by a comma. Only one of these tests should pass}

\item{incorrect_msg}{chracter string to override message generated by the failing test block.}

\item{choose_feedback}{if all tests fail, selects the feedback of the set of tests that should be shown}
}
\description{
Test if one of many test sets pass.
}
\examples{
\dontrun{
# Example solution code:
# a <- 3; b <- 4

# Example SCT
test_or(test_object('a'), test_object('b'))

# Following submissions will all be accepted:
a <- 3; b <- 4
a <- 3
b <- 4
}

}

