% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-output-contains.R
\name{test_output_contains}
\alias{test_output_contains}
\title{Check whether the student printed something to the console}
\usage{
test_output_contains(expr, times = 1, incorrect_msg = NULL,
  env = globalenv())
}
\arguments{
\item{expr}{The expression (as string) for which the output should be in the student's console output.}

\item{times}{How often the expression's output should occur in the student's console}

\item{incorrect_msg}{feeback message in case the output did not contain the expression}

\item{console_output}{The string containing the output printed to the student's console.
The default is DM.console.output which is set on the DataCamp server (automagically).
This means you don't have to specify this argument when writing SCTs for DataCamp.com.}
}
\description{
Function checks whether the student's console contains the output one gets by
evaluating the character string provided in \code{expr} provided to expr.
This function needs refactoring, as all new lines etc are removed.
}
\examples{
\dontrun{
# SCT to test whether student printed numbers 1 to 10
test_output_contains("for(i in 1:10) print(i)")
}
}

