% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-output-contains.R
\name{test_output_contains}
\alias{test_output_contains}
\title{Check whether the student printed something to the console}
\usage{
test_output_contains(expr, times = 1, console_output = get_student_output(),
  incorrect_msg = NULL)
}
\arguments{
\item{expr}{The expression (as string) for which the output should be in the student's console.}

\item{times}{How often the expression's output should occur in the student's console}

\item{console_output}{The string containing the output printed to the student's console.
The default is DM.console.output which is set on the DataCamp server (automagically).
This means you don't have to specify this argument when writing SCTs for DataCamp.com.}

\item{incorrect_msg}{feeback message in case the output did not contain the expression}
}
\description{
Function checks whether the student's console contains the output one gets by
evaluating the character provided to expr. It returns TRUE when the student's
console indeed contains the expected output and FALSE otherwise. This function is
a testwhat version from an earlier datacampSCT function. (copied into this package to reduce dependency)
}
\details{
This test is implemented using \code{\link{test_what}}.
}
\examples{
\dontrun{
# Suppose the student has to write a loop that prints the numbers 1 up to 10.
# Add the following line to the SCT:
test_output_contains("for(i in 1:10) print(i)")
}
}

