% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-output-regex.R
\name{test_output_regex}
\alias{test_output_regex}
\title{Check whether the student output contains a pattern}
\usage{
test_output_regex(pattern, fixed = FALSE, times = 1, incorrect_msg = NULL)
}
\arguments{
\item{pattern}{The pattern to check for in the console output.}

\item{fixed}{logical. if \code{TRUE}, \code{pattern} is matched to the output as is, instead of as regular expression.}

\item{times}{How often the pattern should be matched}

\item{incorrect_msg}{feeback message in case the pattern was not found in the console output}
}
\description{
Check whether the student output contains a pattern
}

