% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-rmd-group.R
\name{test_rmd_group}
\alias{test_rmd_group}
\title{Test a single R Markdown file group (R Markdown exercises)}
\usage{
test_rmd_group(group_number, code, env = parent.frame())
}
\arguments{
\item{group_number}{Number of the group.}

\item{code}{SCT code to test the group (in curly braces)}

\item{env}{The environment in which the code should be tested.}

\item{student_code}{the entire code that has been submitted by the student.}

\item{solution_code}{the entire solution code.}
}
\description{
Test a single R Markdown file group (R Markdown exercises) with arbitrary testwhat functions.
This test is implemented using \code{\link{test_what}}.
}

