% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-regex-old.R
\name{test_student_typed}
\alias{test_student_typed}
\title{Test student's submission as text}
\usage{
test_student_typed(strings, fixed = TRUE, times = 1, not_typed_msg = NULL)
}
\arguments{
\item{strings}{A set of strings that should be available in the student code.}

\item{fixed}{if TRUE, strings are treated literally. If FALSE, strings are treated as regex patterns.}

\item{times}{how often should any of the strings be matched?}

\item{not_typed_msg}{Feedback message in case the student did not type the string.}
}
\description{
Test whether a student typed something in his submission.
Some basic string formatting is performed to allow for different 
ways of saying the same things (removing spaces, changing single 
quotes to double quotes, changing TRUE to T ...).
}
\details{
Using this function should be a last resort, as there are myriad ways of
solving the same problems in R!
}
\examples{
\dontrun{
# Example solution code: TRUE & FALSE

# SCT to test this as a string (both T & F and F & T should be accepted)
test_student_typed(c("TRUE & FALSE", "FALSE & TRUE"))
}

}

