% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-sufficient-length.R
\name{test_sufficient_length}
\alias{test_sufficient_length}
\title{Test sufficient list length.}
\usage{
test_sufficient_length(stud, index, incorrect_number_of_calls_msg = NULL)
}
\arguments{
\item{stud}{list of student commands}

\item{index}{index to fetch from these commands}

\item{incorrect_number_of_calls_msg}{message in case the length of stud is not sufficient.}
}
\description{
Is only used by additional addon testwhat packages; do not use this
function inside an SCT.
}

