% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-wd.R
\name{test_wd}
\alias{test_wd}
\alias{check_wd}
\title{Check whether a file exists}
\usage{
check_wd(state, path, missing_msg = NULL)
}
\arguments{
\item{state}{the state to start from}

\item{path}{Path to the file you want to check}

\item{missing_msg}{Custom feedback message in case the file is missing}

\item{incorrect_msg}{Custom feedback message in case the file does not exist}
}
\description{
Check whether a file exists
}
\examples{
\dontrun{
# Example 1 solution code:
# write("hello", file = "test.txt")

# SCT
ex() \%>\% check_wd("test.txt")
}
}
