% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-what.R
\name{test_what}
\alias{test_what}
\title{Expectation wrapper}
\usage{
test_what(code, feedback)
}
\arguments{
\item{code}{The expectation that should be wrapped}

\item{feedback}{A character string with feedback when the expection is not met 
OR a list object, containing multiple pieces of information.}
}
\description{
This function wraps around an expect_... function. When the expectation fails to be
met, the feedback message is sent to the reporter.
}

