% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test-while-loop.R
\name{test_while_loop}
\alias{test_while_loop}
\title{Check whether the student wrote a while loop correctly}
\usage{
test_while_loop(index = 1, cond_test = NULL, expr_test = NULL,
  student_code = get_student_code(), solution_code = get_solution_code(),
  not_found_msg = NULL, env = parent.frame())
}
\arguments{
\item{index}{The index of the while loop to check.}

\item{cond_test}{SCT to perform on the condition part of the while loop}

\item{expr_test}{SCT to perform on the expression part of the while loop}

\item{student_code}{character string containing the student's code.}

\item{solution_code}{character string containing the sample solution code.}

\item{not_found_msg}{Message in case the while loop (at given index) is not found.}

\item{env}{Environment in which to perform all these SCTs}
}
\description{
Check whether the student wrote a while loop correctly
}

