% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test-yaml-header.R
\name{test_yaml_header}
\alias{test_yaml_header}
\title{Test yaml header (Markdown)}
\usage{
test_yaml_header(options = NULL, check_equality = TRUE,
  allow_extra = TRUE, not_called_msg = NULL, incorrect_msg = NULL)
}
\arguments{
\item{options}{Set of options. Embedded options have to be specified using the dot notation.}

\item{check_equality}{whether or not to actually check the value assigned to the option (default TRUE)}

\item{allow_extra}{whether or not the definition of additional options is accepted (default TRUE)}

\item{not_called_msg}{feedback message if option was not specified (optional but recommended)}

\item{incorrect_msg}{feedback message if option was incorrectly set (optional but recommended)}
}
\description{
Test whether the student specified the correct options in the yaml header (for
R Markdown exercises). This test should be called outside an test_rmd_group call.
}

