% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testwhat.R
\docType{package}
\name{testwhat}
\alias{testwhat}
\alias{testwhat-package}
\title{Easily write submission correctness tests}
\description{
This package makes it easy for teachers to test students' code submissions
for R exercises, e.g., for interactive R courses on \url{www.DataCamp.com}.
}
\note{
\pkg{testwhat} is forked from package \pkg{testthat} such that
teachers can write their tests using a familiar framework.
}
\examples{
library("testwhat")
a <- 9
expect_that(a, is_less_than(10))
expect_less_than(a, 10)
}
\references{
\url{www.DataCamp.com}
}

