% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testwhat.R
\docType{package}
\name{testwhat}
\alias{testwhat}
\alias{testwhat-package}
\title{Easily write submission correctness tests for DataCamp}
\description{
This package makes it easy for teachers to test students' code submissions
for R exercises, e.g., for interactive R courses on \url{www.DataCamp.com}.
}
\details{
This package contains a bunch of high-level functions to test objects, function
calls, output, function definition, control structures and many more.

For more information, see the testwhat vignette.
}
\note{
\pkg{testwhat} is a wrapper around \pkg{testthat} such that
teachers can write their tests using a familiar framework.
}
\references{
\url{www.DataCamp.com}
}

