% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testwhat.R
\docType{package}
\name{testwhat}
\alias{testwhat}
\alias{testwhat-package}
\title{Easily write submission correctness tests for DataCamp}
\description{
This package makes it easy for teachers to test students' code submissions
for R exercises, e.g., for interactive R courses on \url{www.DataCamp.com}.
}
\details{
This package contains a bunch of high-level functions to test objects, function
calls, output, function definition, control structures and many more.

Learn more about testwhat on the wiki: \url{www.github.com/datacamp/testwhat/wiki}.
}
\references{
\url{www.datacamp.com}
}

