% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{tw}
\alias{tw}
\title{"singleton" object to access all information needed by \code{\link{test_what}} functions.}
\format{An object of class \code{list} of length 4.}
\usage{
tw
}
\description{
"singleton" object to access all information needed by \code{\link{test_what}} functions.
}
\keyword{datasets}

